modified modals/portals for less code
