+++ mesoboard-app/src/components/modals/Backdrop.js
@@ -1,0 +1,5 @@
+const Backdrop = ({ onClose, classes }) =>
+  <div className={classes.backdrop} onClick={onClose} />
+
+
+export default Backdrop
