+++ mesoboard-app/src/components/modals/ModalOverlay.js
@@ -1,0 +1,9 @@
+const ModalOverlay = ({ children, classes }) => {
+  return <div className={classes.modal}>
+    <div className={classes.content}>
+      {children}
+    </div>
+  </div>
+}
+
+export default ModalOverlay
