--- mesoboard-app/src/components/modals/NavDrawerModal/NavDrawer.css
@@ -1,3 +1,0 @@
-.navDrawer {
-  
-}
