--- mesoboard-app/src/content/manager/ScheduleManager/ScheduleEdit.js
@@ -1,12 +1,0 @@
-import React from 'react'
-
-const ScheduleEdit = () => {
-  // implement and handle the portal -> ScheduleEditModal component
-  return (
-    <div>
-
-    </div>
-  )
-}
-
-export default ScheduleEdit
